requirements.txt
asgiref==3.2.3
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
decorator==4.4.1
Django==2.2.10
django-appconf==1.0.3
django-compressor==2.4
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.6
django-libsass==0.8
django-sass-processor==0.2.6
djangorestframework==3.11.0
gunicorn==20.0.4
httpie==2.0.0
idna==2.8
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
libsass==0.19.4
parso==0.5.2
pickleshare==0.7.5
prompt-toolkit==3.0.2
psycopg2==2.8.4
Pygments==2.5.2
pytz==2019.3
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
whitenoise==5.0.1
wincertstore==0.2
